+++ R/do_backup.R
@@ -14,23 +14,22 @@
 #' If several backups have been accumulated in a common folder, this function is
 #' able to recognize the newest one.
 #'
-#' @param dbname Character value indicating the name of the target PostgreSQL
-#'     database.
+#' @param dbname,user Character value indicating the name of the target
+#'     PostgreSQL database and user.
 #' @param filename Character value, the name of the backup file. If not provided,
 #'     the function guess the database's name and a time stamp, where the newest
 #'     backup, if many in the same folder, will be used.
 #' @param filepath Character value indicating the path to the folder, where the
 #'     backup file will be restored. It have to be a path relative to the
 #'     working directory.
+#' @param f_timestamp Character value indicating the format of the timestamp
+#'     applied to the backup's name (passed as format to [strptime()]). It can
+#'     be cancelled by 'f_timestamp = NULL', then backups may be overwritten.
 #' @param fext Character value indicating the extension used for the backup
 #'     file, including the leading dot. The extension ".backup" is used by
 #'     default.
 #' @param path_psql Character value indicating the system path to PostgreSQL
 #'     binaries.
-#' @param user Character value indicating the database user in PostgreSQL.
-#' @param password Character value, the user's password in PostgreSQL.
-#' @param f_timestamp Character value indicating the format of the timestamp
-#'     applied to the backup's name (see [strptime()]).
 #' @param host Character value, the host name.
 #' @param port Integer value, the port applied for database connection.
 #' @param backup A character value indicating a specific backup file to be used
@@ -42,29 +41,35 @@
 #' @author Miguel Alvarez
 #'
 #' @export
-do_backup <- function(dbname = "", host = "localhost", port = "5432", user = "",
-                      password = "", filepath = ".", filename = dbname,
-                      fext = ".backup", path_psql = "/usr/bin",
-                      f_timestamp = "%Y%m%d", ...) {
-  # Get credentials (only if not provided)
-  if (user == "" | password == "") {
-    cred <- credentials(user = user, password = password)
-    user <- unname(cred["user"])
-    password <- unname(cred["password"])
+do_backup <- function(dbname, user, host = "localhost", port = "5432",
+                      filepath = ".", filename = dbname, f_timestamp = "%Y%m%d",
+                      fext = ".backup", path_psql = "/usr/bin", ...) {
+  # Get credentials or retrieve an error
+  password <- tryCatch(keyring::key_get(service = dbname, username = user),
+    error = function(e) {
+      stop(paste0(
+        "A password for database '", dbname, "' and user '", user,
+        "' is not yet set\n  Use credentials() to set it."
+      ))
+    }
+  )
+  if (is.null(f_timestamp)) {
+    message("Timestamp set to NULL. Previous backups may be overwritten.")
+    new_file <- file.path(filepath, paste0(filename, fext))
+  } else {
+    old_file <- file.path(filepath, sub(fext, "", list.files(
+      path = filepath,
+      pattern = filename
+    )))
+    new_file <- file.path(filepath, paste0(filename, "_", format(
+      Sys.Date(),
+      f_timestamp
+    )))
+    if (length(old_file) > 0) {
+      new_file <- taxlist:::add_suffix(new_file, old_file)
+    }
+    new_file <- paste0(new_file, fext)
   }
-  # Prepare the file
-  old_file <- file.path(filepath, sub(fext, "", list.files(
-    path = filepath,
-    pattern = filename
-  )))
-  new_file <- file.path(filepath, paste0(filename, "_", format(
-    Sys.Date(),
-    f_timestamp
-  )))
-  if (length(old_file) > 0) {
-    new_file <- taxlist:::add_suffix(new_file, old_file)
-  }
-  new_file <- paste0(new_file, fext)
   # Create command
   system(paste(
     paste0(
@@ -82,20 +87,22 @@ do_backup <- function(dbname = "", host = "localhost", port = "5432", user = "",
 #' @rdname do_backup
 #' @aliases do_restore
 #' @export
-do_restore <- function(dbname = "", backup, host = "localhost", port = "5432",
-                       user = "", password = "", filepath = ".",
-                       filename = dbname, fext = ".backup",
+do_restore <- function(dbname, user, backup, host = "localhost", port = "5432",
+                       filepath = ".", filename = dbname, fext = ".backup",
                        path_psql = "/usr/bin", f_timestamp = "%Y%m%d",
                        opts = "--clean", ...) {
-  # Get credentials (only if not provided)
-  if (user == "" | password == "") {
-    cred <- credentials(user = user, password = password)
-    user <- unname(cred["user"])
-    password <- unname(cred["password"])
-  }
+  # Get credentials or retrieve an error
+  password <- tryCatch(keyring::key_get(service = dbname, username = user),
+    error = function(e) {
+      stop(paste0(
+        "A password for database '", dbname, "' and user '", user,
+        "' is not yet set\n  Use credentials() to set it."
+      ))
+    }
+  )
   # Create command
   if (missing(backup)) {
-    backup <- sort_backups(
+    backup <- taxlist::sort_backups(
       name = filename,
       path = filepath,
       date_format = f_timestamp,